rename protocols.geometry.util protocols.geometry.planning
