+++ robot-server/robot_server/robot/calibration/pipette_offset/user_flow.py
@@ -6,7 +6,7 @@ from opentrons.calibration_storage.types import TipLengthCalNotFound
 from opentrons.config import feature_flags as ff
 from opentrons.hardware_control import ThreadManager, CriticalPoint
 from opentrons.protocol_api import labware
-from opentrons.protocols.geometry import util as geometry
+from opentrons.protocols.geometry import deck
 from opentrons.types import Mount, Point, Location
 from robot_server.service.errors import RobotServerError
 from robot_server.service.session.models import CalibrationCommand
@@ -53,7 +53,7 @@ class PipetteOffsetCalibrationUserFlow:
 
         deck_load_name = SHORT_TRASH_DECK if ff.short_fixed_trash() \
             else STANDARD_DECK
-        self._deck = geometry.Deck(load_name=deck_load_name)
+        self._deck = deck.Deck(load_name=deck_load_name)
         self._tip_rack = self._get_tip_rack_lw()
         self._deck[TIP_RACK_SLOT] = self._tip_rack
